Move hierarchical sidebar visibility logic to lib
